+++ RSPiX/Src/ORANGE/Channel/channel.cpp
@@ -39,8 +39,8 @@
 
 // Usefull test code to make sure it will really compile.  If templated stuff
 // isn't actually instantiated, the compiler will miss many (or all) errors.
-RChannel<int16_t> test;
-RChannel<int32_t> test2(RChannel_ArrayOfPtrs);
+RChannel<short> test;
+RChannel<long> test2(RChannel_ArrayOfPtrs);
 
 
 ////////////////////////////////////////////////////////////////////////////////
