+++ net.cpp
@@ -40,7 +40,7 @@ namespace Net
 	{
 #endif
 
-	int32_t	NETNAMESPACE lBandwidthValues[Net::NumBandwidths] =
+	long	NETNAMESPACE lBandwidthValues[Net::NumBandwidths] =
 		{
 		   1400,						// Analog14_4
 		   2400,						// Analog28_8
